requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.3
constantly==23.10.4
contourpy==1.3.3
coverage==7.10.5
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.0
dj-database-url==2.1.0
Django==5.1
django-cors-headers==4.3.1
django-extensions==4.1
django-filter==24.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
django-allauth==65.11.1
fonttools==4.59.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
matplotlib==3.10.5
msgpack==1.1.1
numpy==1.26.4
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==12.0
zope.interface==7.2
